# backend/app/swagger.yaml
openapi: 3.0.0
info:
  title: API Inventario Forestal JBV2
  description: |
    Sistema de gestión de inventario forestal con servicios ecosistémicos.
    
    ## Funcionalidades principales
    - **Autenticación**: Login con JWT.
    - **Usuarios**: CRUD completo (solo administradores).
    - **Especies**: CRUD completo.
    - **Árboles**: CRUD, mapa, subida de imágenes y cálculos automáticos (biomasa, carbono, CO₂e, área basal, volumen).
    - **Atributos dinámicos**: Gestión de campos adicionales para árboles.
    - **Reportes**: 6 endpoints de estadísticas.
    - **Exportación Darwin Core**: Compatible con GBIF y SiB Colombia (JSON y CSV).
    - **NDVI**: Cálculo del Índice de Vegetación de Diferencia Normalizada usando Sentinel-2.
    - **Administración**: Estructura de tablas para gestión de campos.
  version: 2.0.0
  contact:
    name: Soporte JBV2
    email: soporte@jbv2.com
servers:
  - url: http://localhost:8001
    description: Servidor de desarrollo
security:
  - X-UsuarioID: []
  - BearerAuth: []
paths:
...
  /usuarios:
    get:
      summary: Obtener todos los usuarios
      tags: [Usuarios]
      security: [{ X-UsuarioID: [] }]
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    UsuarioID:
                      type: integer
                    Usuario:
                      type: string
                    RolID:
                      type: integer
    post:
      summary: Crear un nuevo usuario
      tags: [Usuarios]
      security: [{ X-UsuarioID: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Usuario, Contrasena]
              properties:
                Usuario:
                  type: string
                Contrasena:
                  type: string
                RolID:
                  type: integer
                  default: 1
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: integer
  /usuarios/{usuario_id}:
    put:
      summary: Actualizar un usuario
      tags: [Usuarios]
      security: [{ X-UsuarioID: [] }]
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Usuario:
                  type: string
                Contrasena:
                  type: string
                RolID:
                  type: integer
      responses:
        '200':
          description: Usuario actualizado
    delete:
      summary: Eliminar un usuario
      tags: [Usuarios]
      security: [{ X-UsuarioID: [] }]
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario eliminado

  # ============================================================
  # ESPECIES
  # ============================================================
...
components:
  securitySchemes:
    X-UsuarioID:
      type: apiKey
      in: header
      name: X-UsuarioID
      description: ID del usuario autenticado (ej. 1). Obténlo al hacer login.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /auth/login